+++ hc/integrations/matrix/client.py
@@ -1,11 +1,10 @@
 from __future__ import annotations
 
-from urllib.parse import quote, urlencode
+from urllib.parse import quote
 
 from django.conf import settings
-from pydantic import BaseModel, Field, ValidationError
-
 from hc.lib import curl
+from pydantic import BaseModel, Field, ValidationError
 
 JOIN_ERRORS = {
     429: "Matrix server returned status 429 (Too Many Requests), please try later.",
@@ -25,9 +24,9 @@ class MatrixJoinResponse(BaseModel):
 def join(alias: str) -> str:
     assert settings.MATRIX_HOMESERVER
     url = settings.MATRIX_HOMESERVER
-    url += "/_matrix/client/r0/join/%s?" % quote(alias)
-    url += urlencode({"access_token": settings.MATRIX_ACCESS_TOKEN})
-    r = curl.post(url)
+    url += "/_matrix/client/v3/join/" + quote(alias)
+    headers = {"Authorization": f"Bearer {settings.MATRIX_ACCESS_TOKEN}"}
+    r = curl.post(url, headers=headers)
     if r.status_code in JOIN_ERRORS:
         raise JoinError(JOIN_ERRORS[r.status_code])